Update faker in composer, update the content the our model and generate fake data for testing purposes. Fix the migrations.

+++ database/migrations/2016_07_05_151442_create_clients_table.php
@@ -17,7 +17,7 @@ class CreateClientsTable extends Migration
             $table->string('name',25);
             $table->string('identity_card',20)->unique();
             $table->smallInteger('age')->nullable()->unsigned();
-            $table->smallInteger('phone')->nullable()->unsigned();
+            $table->integer('phone')->nullable()->unsigned();
             $table->string('address',50)->nullable();
             $table->timestamps();  // With timestamps we create the following fields created_at and updated_at
